move counters to their own struct
